Move the OTA sentinel beside the firmware job types
