+++ test/components/logs-dialog-quiet-serial.test.ts
@@ -8,8 +8,8 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import { call, makeLogsDialog, session } from "./_logs-dialog-env.js";
 
+import { OTA_PORT } from "../../src/api/types/firmware-jobs.js";
 import { switchToOtaLogs } from "../../src/components/logs-dialog/session.js";
-import { OTA_PORT } from "../../src/components/logs-session.js";
 import type { ESPHomeLogsDialog } from "./_logs-dialog-env.js";
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
